requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
tortoise-orm==0.23.0
aerich==0.8.1
SQLAlchemy==2.0.41
asyncpg==0.30.0
aiosqlite==0.20.0
passlib==1.7.4
bcrypt==4.3.0
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.9.1
argon2-cffi==23.1.0
PyJWT==2.10.1
google-genai==1.61.0
httpx>=0.27.0
apscheduler==3.10.4
tzdata>=2024.1
azure-cognitiveservices-speech==1.40.0
